Prune payment processors from session before saving it.
I waas able to replicate the session errors reported in #145 as being due to the Processor objects being stored to the session
and then not unserializing well due to load order. Per comments I don't think there is any benefit in having them there.

The could plausibly cause the other symptoms reported by interfering with the code's efforts to retrieve the
url to forward the user from the session and / or by causing the redirect to fail due to php having already output some noise